created strategy table, updated types, relationhip, sma calculation

+++ TradeBash/src/TradeBash.Infrastructure/DTO/Stock.cs
@@ -8,9 +8,9 @@ namespace TradeBash.Infrastructure.DTO
         
         public string Symbol { get; set; }
         
-        public double Open { get; set; }
+        public decimal Open { get; set; }
         
-        public double Close { get; set; }
+        public decimal Close { get; set; }
         
         public string Label { get; set; }
     }
